requirements.txt
pillow
numpy
scipy
